+++ src/object.h
@@ -60,9 +60,9 @@ PMOD_EXPORT void do_free_object(struct object *o);
 PMOD_EXPORT struct object *debug_clone_object(struct program *p, int args);
 PMOD_EXPORT struct object *fast_clone_object(struct program *p, int args);
 PMOD_EXPORT struct object *parent_clone_object(struct program *p,
-				   struct object *parent,
-				   int parent_identifier,
-				   int args);
+					       struct object *parent,
+					       ptrdiff_t parent_identifier,
+					       int args);
 PMOD_EXPORT struct object *get_master(void);
 PMOD_EXPORT struct object *debug_master(void);
 struct destroy_called_mark;
