+++ src/program.c
@@ -2703,7 +2703,7 @@ int really_low_find_shared_string_identifier(struct pike_string *name,
   return -1;
 }
 
-int low_find_lfun(struct program *p, int lfun)
+int low_find_lfun(struct program *p, ptrdiff_t lfun)
 {
   struct pike_string *lfun_name = findstring(lfun_names[lfun]);
   unsigned int flags = 0;
